requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.7
mcp>=0.9.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.54.0
anthropic>=0.39.0

